Fixed closing service line item bug
