+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/EditContractModal.tsx
@@ -131,6 +131,12 @@ export const EditContractModal = ({
 
         return;
       }
+      if (e.closed) {
+        contractLineItemsStore.closeServiceLineItem(e.metadata.id, contractId);
+
+        return;
+      }
+
       itemStore?.update((prev) => prev);
     });
